backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
python-dotenv==1.0.1
groq==0.5.0
httpx==0.27.0
python-multipart==0.0.9